refactor(beta/basic-table): drop virtualization (BYO), remove vendored internal/types, relocate test-utils to root
- Remove use-virtualization primitive + its 3 test suites; virtualization is now
  bring-your-own (consumer spreads offset/measure props onto Body/Row).
- Delete the vendored internal/ and types/ folders; re-point apply-display-name, events,
  and the base-component/event types to the repo originals (extends the toolkit-revert).
- Move test-utils from beta/basic-table-0.1/test-utils to src/test-utils/dom/basic-table so
  it is generated + exported like every other component (drops the beta test-utils subpath).
- Reword USAGE.md + interface/code comments to reflect BYO virtualization.
