+++ src/beta/basic-table-0.1/__tests__/basic-table.test.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 
-import createWrapper from '../../../../lib/components/beta/basic-table-0.1/test-utils/dom';
+import createWrapper from '../../../../lib/components/test-utils/dom';
 import BasicTable, {
   BasicTableHeader,
   BasicTableHeaderCell,
