+++ src/beta/basic-table-0.1/basic-table-cell/index.tsx
@@ -4,7 +4,7 @@
 import React from 'react';
 
 import { Cell } from '../basic-table/internal';
-import { applyDisplayName } from '../internal/utils/apply-display-name';
+import { applyDisplayName } from '../../../internal/utils/apply-display-name';
 import { BasicTableProps } from '../basic-table/interfaces';
 
 export type BasicTableCellProps = BasicTableProps.CellProps;
