+++ src/beta/basic-table-0.1/basic-table-header-cell/index.tsx
@@ -4,7 +4,7 @@
 import React from 'react';
 
 import { HeaderCell } from '../basic-table/internal';
-import { applyDisplayName } from '../internal/utils/apply-display-name';
+import { applyDisplayName } from '../../../internal/utils/apply-display-name';
 import { BasicTableProps } from '../basic-table/interfaces';
 
 export type BasicTableHeaderCellProps = BasicTableProps.HeaderCellProps;
