+++ src/beta/basic-table-0.1/basic-table/index.tsx
@@ -4,7 +4,7 @@
 import React from 'react';
 
 import useBaseComponent from '../../../internal/hooks/use-base-component';
-import { applyDisplayName } from '../internal/utils/apply-display-name';
+import { applyDisplayName } from '../../../internal/utils/apply-display-name';
 import { BasicTableProps } from './interfaces';
 import { InternalRoot } from './internal';
