--- src/beta/basic-table-0.1/internal/environment.ts
@@ -1,11 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-// Build environment constants.
-export const THEME = 'default';
-export const SYSTEM = 'core';
-export const PACKAGE_SOURCE = 'basic-table-components';
-export const PACKAGE_VERSION = '';
-export const GIT_SHA = '';
-/** Indicates that the current theme is always in visual refresh mode. */
-export const ALWAYS_VISUAL_REFRESH = false;
