--- src/beta/basic-table-0.1/internal/hooks/use-visual-mode.ts
@@ -1,8 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-import { useRuntimeVisualRefresh } from '@cloudscape-design/component-toolkit/internal';
-
-import { ALWAYS_VISUAL_REFRESH } from '../environment';
-
-export const useVisualRefresh = ALWAYS_VISUAL_REFRESH ? () => true : useRuntimeVisualRefresh;
