--- src/beta/basic-table-0.1/internal/utils/apply-display-name.ts
@@ -1,5 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-export function applyDisplayName<T>(component: T, displayName: string): void {
-  (component as any).displayName = displayName;
-}
