--- src/beta/basic-table-0.1/test-utils/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../dist/test-utils",
-    "rootDir": ".",
-    "declaration": true,
-    "tsBuildInfoFile": "../../dist/.cache/test-utils.tsbuildinfo"
-  },
-  "include": ["./", "./types"],
-  "exclude": []
-}
