--- src/beta/basic-table-0.1/use-virtualization/index.ts
@@ -1,13 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-export { useVirtualization } from './use-virtualization';
-export { useColumnVirtualization } from './use-column-window';
-export type {
-  VirtualizationConfig,
-  VirtualizationResult,
-  VirtualizationWindowItem,
-  VirtualizationRowProps,
-  VirtualizationVisibleRange,
-  ColumnVirtualizationConfig,
-  ColumnVirtualizationResult,
-} from './interfaces';
